updated routes

+++ routes/api.php
@@ -21,14 +21,15 @@ Route::post('/users', [UsersController::class, 'createUser']);
 Route::put('/users/{user_id}', [UsersController::class, 'updateUser']);
 Route::put('/users/{user_id}/change_password', [UsersController::class, 'changeUserPassword']);
 Route::delete('/users/{user_id}', [UsersController::class, 'deleteUser']);
+
 Route::get('/users/{user_id}/stuff', [StuffController::class, 'getStuff']);
+Route::get('/users/{user_id}/stuff/{stuff_id}', [PackagesController::class, 'getPackages']);
+
+Route::post('/users/{user_id}/packages', [PackagesController::class, 'createPackage']);
+Route::get('/users/{user_id}/packages/{uuid}', [PackagesController::class, 'getPackageByUUID']);
+Route::put('/users/{user_id}/packages/{uuid}', [PackagesController::class, 'updatePackage']);
+Route::delete('/users/{user_id}/packages/{uuid}', [PackagesController::class, 'deletePackage']);
 
 Route::post('/stuff', [StuffController::class, 'createStuff']);
 Route::put('/stuff/{stuff_id}', [StuffController::class, 'updateStuff']);
 Route::delete('/stuff/{stuff_id}', [StuffController::class, 'deleteStuff']);
-
-Route::get('/users/{user_id}/stuff/{stuff_id}', [PackagesController::class, 'getPackages']);
-Route::get('/users/{user_id}/packages/{uuid}', [PackagesController::class, 'getPackageByUUID']);
-Route::post('/packages', [PackagesController::class, 'createPackage']);
-Route::put('/packages/{uuid}', [PackagesController::class, 'updatePackage']);
-Route::delete('/packages/{uuid}', [PackagesController::class, 'deletePackage']);
